refactor: [노트 모아보기] useInfiniteNotes의 return data를 flat하게 가공
