+++ src/views/notes/empty-notes/EmptyNotes.tsx
@@ -1,0 +1,7 @@
+export default function EmptyNotes() {
+  return (
+    <h1 className="mt-[300px] self-center text-sm font-normal text-slate-500">
+      아직 등록된 노트가 없어요
+    </h1>
+  );
+}
